move: users 스키마 이동
